Add support for using proxies
The etcd proxies are a nice feature of version 2.x, and they just act as
a proxy between the actual cluster and the client. Since there can be
cases (like multi-dc or as a way to add high-availability to the client
in case of a node failure) where proxies will be used, we allow the user
to add a flag to the etcd client so that proxies can be used - in
particular, if you decide to use proxies AND our own reconnection
feature you will be able to connect only to the list of servers that you
passed onto etcd at startup.

This is clearly a downgrade compared to what python-etcd normally
does (when it figures out by itself not only the failure of a machine,
but it will also refresh the list of servers based on the response from
the one still active server we found), but has I stated, complex
environment may need to use proxies.

This commit takes care of that.